+++ docs/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+### Changed
+
+- Return `FailedOperation` instead of raising exceptions in `.compute()` methods.
+- Use `TCFEKeywords` enums throughout code.
+- Collect `tc.out`, if possible, when computation fails.
+
 ## [0.10.1]
 
 ### Changed